chore(wizard-refactor): extract tools/mark_complete
Pinned by 4 dispatcher tests (no draftPlaybookId / playbook missing /
no curriculum modules / happy path). 28/28 green.
